refactor: Move Logo component to common directory
