+++ client/src/components/common/Logo/index.stories.tsx
