+++ client/src/components/common/Logo/index.tsx
