+++ client/src/components/common/Logo/types.ts
